offline_representations/requirements.txt
atari-py==0.2.6
dm-acme==0.2.0
dm-env==1.4
dm-reverb==0.2.0
dm-sonnet==2.0.0
dm-tree==0.1.6
dopamine-rl==3.1.2
frozendict==1.2
future==0.18.2
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gym==0.18.0
h5py==2.10.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
mypy-extensions==0.4.3
numpy==1.20.2
PyDrive==1.3.1
pygame==2.0.1
PyYAML==5.4.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.1
toml==0.10.2
trfl==1.1.0
wandb==0.10.26
Werkzeug==1.0.1
